backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1
mysql-connector-python==9.0.0
PyMySQL==1.1.0
cryptography==42.0.5
bcrypt==4.1.2
marshmallow==3.21.1
python-dotenv==1.0.1
razorpay==1.4.1
reportlab==4.1.0
sendgrid==6.11.0
gunicorn==22.0.0
Flask-Limiter==3.7.0
APScheduler==3.10.4
pandas==2.2.2
